+++ src/components/input/cardPassword/input.tsx
@@ -1,0 +1,13 @@
+const CardPasswordInput = () => {
+  return (
+    <div className="input-container">
+      <span className="input-title">카드 비밀번호</span>
+      <input className="input-basic w-15" type="password" />
+      <input className="input-basic w-15" type="password" />
+      <input className="input-basic w-15" type="password" />
+      <input className="input-basic w-15" type="password" />
+    </div>
+  );
+};
+
+export default CardPasswordInput;
