+++ src/components/input/cardSecurityNumber/input.tsx
@@ -1,0 +1,10 @@
+const CardCVCInput = () => {
+  return (
+    <div className="input-container">
+      <span className="input-title">보안코드(CVC/CVV)</span>
+      <input className="input-basic w-25" type="password" />
+    </div>
+  );
+};
+
+export default CardCVCInput;
